# 🚀 Performance Optimization: EasyOCR Singleton + Parallel Processing